chore: normalize generated code
