+++ packages/nx-playwright/executors.json
@@ -1,10 +1,10 @@
 {
   "$schema": "http://json-schema.org/schema",
   "executors": {
-    "build": {
-      "implementation": "./src/executors/build/executor",
-      "schema": "./src/executors/build/schema.json",
-      "description": "build executor"
+    "playwright": {
+      "implementation": "./src/executors/playwright/executor",
+      "schema": "./src/executors/playwright/schema.json",
+      "description": "playwright executor"
     }
   }
 }
