+++ packages/nx-playwright/generators.json
@@ -3,10 +3,10 @@
   "name": "nx-playwright",
   "version": "0.0.1",
   "generators": {
-    "nx-playwright": {
-      "factory": "./src/generators/nx-playwright/generator",
-      "schema": "./src/generators/nx-playwright/schema.json",
-      "description": "nx-playwright generator"
+    "project": {
+      "factory": "./src/generators/project/generator",
+      "schema": "./src/generators/project/schema.json",
+      "description": "playwright project generator"
     }
   }
 }
