+++ packages/nx-playwright/src/executors/playwright/executor.ts
@@ -1,0 +1,8 @@
+import { PlaywrightExecutorSchema } from './schema';
+
+export default async function runExecutor(options: PlaywrightExecutorSchema) {
+  console.log('Executor ran for Build', options);
+  return {
+    success: true,
+  };
+}
