+++ packages/nx-playwright/src/executors/playwright/schema.d.ts
@@ -1,0 +1,1 @@
+export interface PlaywrightExecutorSchema {} // eslint-disable-line
