+++ packages/nx-playwright/src/executors/playwright/schema.json
@@ -2,7 +2,7 @@
   "$schema": "http://json-schema.org/schema",
   "version": 2,
   "cli": "nx",
-  "title": "Build executor",
+  "title": "Playwright executor",
   "description": "",
   "type": "object",
   "properties": {},
