--- packages/nx-playwright/src/generators/nx-playwright/schema.d.ts
@@ -1,5 +1,0 @@
-export interface NxPlaywrightGeneratorSchema {
-  name: string;
-  tags?: string;
-  directory?: string;
-}
