+++ packages/nx-playwright/src/generators/project/files/src/index.ts__template__
