+++ packages/nx-playwright/src/generators/project/generator.spec.ts
@@ -1,11 +1,11 @@
 import { readProjectConfiguration, Tree } from '@nrwl/devkit';
 import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
 import generator from './generator';
-import { NxPlaywrightGeneratorSchema } from './schema';
+import { PlaywrightGeneratorSchema } from './schema';
 
 describe('nx-playwright generator', () => {
   let appTree: Tree;
-  const options: NxPlaywrightGeneratorSchema = { name: 'test' };
+  const options: PlaywrightGeneratorSchema = { name: 'test' };
 
   beforeEach(() => {
     appTree = createTreeWithEmptyWorkspace({ layout: 'apps-libs' });
