+++ packages/nx-playwright/src/generators/project/schema.d.ts
@@ -1,0 +1,11 @@
+import { NormalizedOptions } from '@ns3/nx-core';
+
+export interface PlaywrightGeneratorSchema {
+  name: string;
+  tags?: string;
+  directory?: string;
+}
+
+export interface PlaywrightGeneratorNormalizedSchema
+  extends PlaywrightGeneratorSchema,
+    NormalizedOptions {}
