+++ packages/nx-playwright/src/generators/project/schema.json
@@ -1,8 +1,8 @@
 {
   "$schema": "http://json-schema.org/schema",
   "cli": "nx",
-  "$id": "NxPlaywright",
-  "title": "",
+  "$id": "NxPlaywrightProject",
+  "title": "Add a Playwright Project",
   "type": "object",
   "properties": {
     "name": {
